work on version push

+++ builtin/providers/vsphere/resource_vsphere_virtual_machine.go
@@ -32,10 +32,10 @@ type networkInterface struct {
 	deviceName       string
 	label            string
 	ipv4Address      string
-	ipv4PrefixLength int
+	ipv4PrefixLength int32
 	ipv4Gateway      string
 	ipv6Address      string
-	ipv6PrefixLength int
+	ipv6PrefixLength int32
 	ipv6Gateway      string
 	adapterType      string // TODO: Make "adapter_type" argument
 }
@@ -72,7 +72,7 @@ type virtualMachine struct {
 	cluster               string
 	resourcePool          string
 	datastore             string
-	vcpu                  int
+	vcpu                  int32
 	memoryMb              int64
 	memoryAllocation      memoryAllocation
 	template              string
@@ -423,7 +423,7 @@ func resourceVSphereVirtualMachineUpdate(d *schema.ResourceData, meta interface{
 	configSpec := types.VirtualMachineConfigSpec{}
 
 	if d.HasChange("vcpu") {
-		configSpec.NumCPUs = d.Get("vcpu").(int)
+		configSpec.NumCPUs = d.Get("vcpu").(int32)
 		hasChanges = true
 		rebootRequired = true
 	}
@@ -505,7 +505,7 @@ func resourceVSphereVirtualMachineCreate(d *schema.ResourceData, meta interface{
 
 	vm := virtualMachine{
 		name:     d.Get("name").(string),
-		vcpu:     d.Get("vcpu").(int),
+		vcpu:     d.Get("vcpu").(int32),
 		memoryMb: int64(d.Get("memory").(int)),
 		memoryAllocation: memoryAllocation{
 			reservation: int64(d.Get("memory_reservation").(int)),
@@ -587,7 +587,7 @@ func resourceVSphereVirtualMachineCreate(d *schema.ResourceData, meta interface{
 				if ip != nil {
 					mask := net.IPv4Mask(ip[0], ip[1], ip[2], ip[3])
 					pl, _ := mask.Size()
-					networks[i].ipv4PrefixLength = pl
+					networks[i].ipv4PrefixLength = int32(pl)
 				} else {
 					return fmt.Errorf("subnet_mask parameter is invalid.")
 				}
@@ -595,7 +595,7 @@ func resourceVSphereVirtualMachineCreate(d *schema.ResourceData, meta interface{
 			if v, ok := network["ipv4_address"].(string); ok && v != "" {
 				networks[i].ipv4Address = v
 			}
-			if v, ok := network["ipv4_prefix_length"].(int); ok && v != 0 {
+			if v, ok := network["ipv4_prefix_length"].(int32); ok && v != 0 {
 				networks[i].ipv4PrefixLength = v
 			}
 			if v, ok := network["ipv4_gateway"].(string); ok && v != "" {
@@ -604,7 +604,7 @@ func resourceVSphereVirtualMachineCreate(d *schema.ResourceData, meta interface{
 			if v, ok := network["ipv6_address"].(string); ok && v != "" {
 				networks[i].ipv6Address = v
 			}
-			if v, ok := network["ipv6_prefix_length"].(int); ok && v != 0 {
+			if v, ok := network["ipv6_prefix_length"].(int32); ok && v != 0 {
 				networks[i].ipv6PrefixLength = v
 			}
 			if v, ok := network["ipv6_gateway"].(string); ok && v != "" {
@@ -1046,7 +1046,7 @@ func buildNetworkDevice(f *find.Finder, label, adapterType string) (*types.Virtu
 
 // buildVMRelocateSpec builds VirtualMachineRelocateSpec to set a place for a new VirtualMachine.
 func buildVMRelocateSpec(rp *object.ResourcePool, ds *object.Datastore, vm *object.VirtualMachine, linkedClone bool, initType string) (types.VirtualMachineRelocateSpec, error) {
-	var key int
+	var key int32
 	var moveType string
 	if linkedClone {
 		moveType = "createNewChildDiskBacking"
@@ -1139,7 +1139,7 @@ func buildStoragePlacementSpecClone(c *govmomi.Client, f *object.DatacenterFolde
 		return types.StoragePlacementSpec{}
 	}
 
-	var key int
+	var key int32
 	for _, d := range devices.SelectByType((*types.VirtualDisk)(nil)) {
 		key = d.GetVirtualDevice().Key
 		log.Printf("[DEBUG] findDatastore: virtual devices: %#v\n", d.GetVirtualDevice())
@@ -1504,7 +1504,7 @@ func (vm *virtualMachine) deployVirtualMachine(c *govmomi.Client) error {
 			if network.ipv4PrefixLength == 0 {
 				return fmt.Errorf("Error: ipv4_prefix_length argument is empty.")
 			}
-			m := net.CIDRMask(network.ipv4PrefixLength, 32)
+			m := net.CIDRMask(int(network.ipv4PrefixLength), 32)
 			sm := net.IPv4(m[0], m[1], m[2], m[3])
 			subnetMask := sm.String()
 			log.Printf("[DEBUG] ipv4 gateway: %v\n", network.ipv4Gateway)
@@ -1592,7 +1592,7 @@ func (vm *virtualMachine) deployVirtualMachine(c *govmomi.Client) error {
 		guiUnattended := types.CustomizationGuiUnattended{
 			AutoLogon:      false,
 			AutoLogonCount: 1,
-			TimeZone:       timeZone,
+			TimeZone:       int32(timeZone),
 		}
 
 		customIdentification := types.CustomizationIdentification{}
